% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bass.R
\name{apply_adsr}
\alias{apply_adsr}
\title{Apply ADSR envelope to samples}
\usage{
apply_adsr(samples, sample_rate, attack, decay, sustain_level, release)
}
\arguments{
\item{samples}{Numeric vector of audio samples}

\item{sample_rate}{Sample rate in Hz}

\item{attack}{Attack time in seconds}

\item{decay}{Decay time in seconds}

\item{sustain_level}{Sustain level from 0.0 to 1.0}

\item{release}{Release time in seconds}
}
\value{
Numeric vector of enveloped samples
}
\description{
Internal helper that applies an Attack-Decay-Sustain-Release
envelope to a sample vector.
}
\keyword{internal}
