% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{apply_delay}
\alias{apply_delay}
\title{Apply Delay to Audio Buffer}
\usage{
apply_delay(wave, delay_params)
}
\arguments{
\item{wave}{tuneR Wave object}

\item{delay_params}{List from calculate_delay_params()}
}
\value{
Wave object with delay applied
}
\description{
Simple delay line with feedback.
}
\keyword{internal}
