% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_theory.R
\docType{data}
\name{CHORD_FORMULAS}
\alias{CHORD_FORMULAS}
\title{Chord interval formulas (semitones from root)}
\format{
An object of class \code{list} of length 8.
}
\usage{
CHORD_FORMULAS
}
\description{
Named list of chord formulas used in deep house production.
Each formula is a numeric vector of semitone offsets from the root note.
}
\keyword{datasets}
