% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_model.R
\name{code_model}
\alias{code_model}
\title{CodeModel R6 Class}
\description{
R6 class for immutable representation of analyzed R code.
Stores extracted features (functions, variables, calls) and computed
musical parameters for audio generation.
}
