% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\docType{data}
\name{DEFAULT_MIX_LEVELS}
\alias{DEFAULT_MIX_LEVELS}
\title{Default mix levels for each instrument type}
\format{
An object of class \code{list} of length 8.
}
\usage{
DEFAULT_MIX_LEVELS
}
\description{
Named list of level multipliers (0-1) for each instrument.
These values are tuned for a balanced deep house mix.
}
\keyword{datasets}
