% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{EXPORT_BIT}
\alias{EXPORT_BIT}
\title{Export bit depth (16-bit for CD quality WAV export)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
EXPORT_BIT
}
\description{
Export bit depth (16-bit for CD quality WAV export)
}
\keyword{datasets}
