% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drums.R
\name{get_drum_kit_cached}
\alias{get_drum_kit_cached}
\title{Get cached drum kit}
\usage{
get_drum_kit_cached()
}
\value{
Named list of drum samples
}
\description{
Internal function that loads and caches the drum kit.
}
\keyword{internal}
