% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_generator.R
\name{hash_to_progression}
\alias{hash_to_progression}
\title{Generate a chord progression from a name}
\usage{
hash_to_progression(name, length = 4)
}
\arguments{
\item{name}{Character string to hash}

\item{length}{Integer number of chords in progression. Default 4.}
}
\value{
List of chord specs, each with root (scale degree) and type
}
\description{
Hashes a name to deterministically generate a chord progression
for the track.
}
\details{
The progression uses common deep house chord types: min7, maj7, dom7, min9.
Root notes are expressed as scale degrees (1-7) to be relative to the key.
}
\examples{
prog <- hash_to_progression("clean_data")
# Returns list of chord specs like:
# list(list(root = 1, type = "min7"), list(root = 4, type = "maj7"), ...)

}
