% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_theory.R
\docType{data}
\name{HOUSE_KEYS}
\alias{HOUSE_KEYS}
\title{Common deep house root notes (MIDI numbers)}
\format{
An object of class \code{list} of length 5.
}
\usage{
HOUSE_KEYS
}
\description{
Named list of root notes commonly used in deep house.
Values are MIDI note numbers for the third octave (bass range).
}
\keyword{datasets}
