% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_functions_to_parts}
\alias{map_functions_to_parts}
\title{Map Functions to Parts}
\usage{
map_functions_to_parts(functions)
}
\arguments{
\item{functions}{Character vector of function names}
}
\value{
Named list mapping function names to parts (1-4)
}
\description{
Hash each function name to determine part assignment (1-4).
Uses MD5 hash for determinism.
}
\keyword{internal}
