% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_variables_to_instruments}
\alias{map_variables_to_instruments}
\title{Map Variables to Instruments}
\usage{
map_variables_to_instruments(variables)
}
\arguments{
\item{variables}{Character vector of variable names}
}
\value{
Named list mapping variable names to instruments (1-5)
}
\description{
Hash each variable name to determine instrument assignment (1-5).
Uses MD5 hash for determinism.
}
\keyword{internal}
