% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PCM_MODE}
\alias{PCM_MODE}
\title{PCM mode (FALSE = IEEE float for 32-bit processing)}
\format{
An object of class \code{logical} of length 1.
}
\usage{
PCM_MODE
}
\description{
PCM mode (FALSE = IEEE float for 32-bit processing)
}
\keyword{datasets}
