% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_theory.R
\name{raver_build_chord}
\alias{raver_build_chord}
\title{Build a chord from a root note}
\usage{
raver_build_chord(root_midi, chord_type = "min9")
}
\arguments{
\item{root_midi}{Integer MIDI note number for the root}

\item{chord_type}{Character name of chord type from CHORD_FORMULAS,
or numeric vector of semitone intervals. Default "min9".}
}
\value{
Integer vector of MIDI note numbers in the chord
}
\description{
Builds a chord from a root MIDI note using a specified formula.
Returns MIDI note numbers for the chord.
}
\examples{
raver_build_chord(48, "min9")   # Cmin9 = C, Eb, G, Bb, D
raver_build_chord(57, "min7")   # Amin7 = A, C, E, G

}
