% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_theory.R
\name{raver_build_scale}
\alias{raver_build_scale}
\title{Build a scale from a root note}
\usage{
raver_build_scale(root_midi, pattern = "natural_minor", octaves = 2)
}
\arguments{
\item{root_midi}{Integer MIDI note number for the root}

\item{pattern}{Character name of scale pattern from SCALE_PATTERNS,
or numeric vector of semitone intervals. Default "natural_minor".}

\item{octaves}{Integer number of octaves to span. Default 2.}
}
\value{
Integer vector of MIDI note numbers in the scale
}
\description{
Builds a scale from a root MIDI note using a specified pattern.
Returns MIDI note numbers spanning the requested number of octaves.
}
\examples{
raver_build_scale(57, "natural_minor", 2)  # A minor, 2 octaves
raver_build_scale(48, "major", 1)          # C major, 1 octave

}
