% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_theory.R
\name{raver_get_chord_freqs}
\alias{raver_get_chord_freqs}
\title{Get chord frequencies}
\usage{
raver_get_chord_freqs(root_midi, chord_type = "min9")
}
\arguments{
\item{root_midi}{Integer MIDI note number for the root}

\item{chord_type}{Character name of chord type from CHORD_FORMULAS.
Default "min9".}
}
\value{
Numeric vector of frequencies in Hz
}
\description{
Convenience function that builds a chord and returns
frequencies in Hz instead of MIDI note numbers.
}
\examples{
raver_get_chord_freqs(48, "min9")  # Cmin9 frequencies

}
