% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_theory.R
\name{raver_invert_chord}
\alias{raver_invert_chord}
\title{Invert a chord for voice leading}
\usage{
raver_invert_chord(chord_notes, inversion = 0)
}
\arguments{
\item{chord_notes}{Integer vector of MIDI note numbers}

\item{inversion}{Integer inversion number.
0 = root position, 1 = first inversion, 2 = second inversion, etc.
Wraps around if inversion > chord length.}
}
\value{
Integer vector of MIDI note numbers in the inverted chord
}
\description{
Creates a chord inversion by moving the bottom N notes
up an octave. Used for smooth voice leading between chords.
}
\examples{
raver_invert_chord(c(48, 51, 55), 1)  # Move C up: Eb, G, C
raver_invert_chord(c(48, 51, 55), 2)  # Move C, Eb up: G, C, Eb

}
