% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_theory.R
\name{raver_note_to_midi}
\alias{raver_note_to_midi}
\title{Convert note name and octave to MIDI note number}
\usage{
raver_note_to_midi(note, octave)
}
\arguments{
\item{note}{Character note name. Supports:
\itemize{
\item Natural notes: C, D, E, F, G, A, B
\item Sharps: C#, Cs, D#, Ds, F#, Fs, G#, Gs, A#, As
\item Flats: Db, Df, Eb, Ef, Gb, Gf, Ab, Af, Bb, Bf
}}

\item{octave}{Integer octave number (0-10 typical range)}
}
\value{
Integer MIDI note number
}
\description{
Converts a note name (e.g., "C", "F#", "Bb") and octave
to its corresponding MIDI note number.
}
\details{
MIDI note numbering:
\itemize{
\item C-1 = 0, C0 = 12, C4 = 60 (middle C), A4 = 69
\item Formula: (octave + 1) * 12 + note_offset + modifier
}
}
\examples{
raver_note_to_midi("C", 4)   # 60 (middle C)
raver_note_to_midi("A", 4)   # 69 (concert A)
raver_note_to_midi("C#", 4)  # 61
raver_note_to_midi("Bb", 4)  # 70

}
