% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_loader.R
\name{resample_wave}
\alias{resample_wave}
\title{Resample a Wave object to target sample rate}
\usage{
resample_wave(wave, target_rate = SAMPLE_RATE)
}
\arguments{
\item{wave}{A Wave object}

\item{target_rate}{Target sample rate (default: SAMPLE_RATE = 44100)}
}
\value{
Wave object at target sample rate
}
\description{
Uses linear interpolation to convert sample rate.
}
\keyword{internal}
