% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composer.R
\name{reset_rng_state}
\alias{reset_rng_state}
\title{Reset RNG state}
\usage{
reset_rng_state()
}
\description{
Resets R's random state after composition to prevent
composition from affecting user's random state.
}
\keyword{internal}
