% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_theory.R
\docType{data}
\name{SCALE_PATTERNS}
\alias{SCALE_PATTERNS}
\title{Scale interval patterns (semitones from root)}
\format{
An object of class \code{list} of length 6.
}
\usage{
SCALE_PATTERNS
}
\description{
Named list of scale patterns used in deep house production.
Each pattern is a numeric vector of semitone offsets from the root note.
}
\keyword{datasets}
