% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composer.R
\name{select_bpm_for_genre}
\alias{select_bpm_for_genre}
\title{Select BPM for Genre}
\usage{
select_bpm_for_genre(code_model, genre = "deep_house")
}
\arguments{
\item{code_model}{A CodeModel object}

\item{genre}{Character genre name}
}
\value{
Integer BPM within genre range
}
\description{
Deterministically selects BPM within genre-appropriate range.
}
\keyword{internal}
