% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_loader.R
\name{stereo_to_mono}
\alias{stereo_to_mono}
\title{Convert stereo Wave to mono}
\usage{
stereo_to_mono(wave)
}
\arguments{
\item{wave}{A Wave object}
}
\value{
Mono Wave object
}
\description{
Averages left and right channels.
}
\keyword{internal}
