% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{BIT_DEPTH}
\alias{BIT_DEPTH}
\title{Internal processing bit depth (32-bit float)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
BIT_DEPTH
}
\description{
Internal processing bit depth (32-bit float)
}
\examples{
# Check the internal bit depth
BIT_DEPTH

}
\keyword{datasets}
