% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SAMPLE_RATE}
\alias{SAMPLE_RATE}
\title{Sample rate in Hz (CD quality)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
SAMPLE_RATE
}
\description{
Sample rate in Hz (CD quality)
}
\examples{
# Use SAMPLE_RATE when creating Wave objects
SAMPLE_RATE

}
\keyword{datasets}
